requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
airtable==0.4.8
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
discord-webhook==0.17.0
discord.py==2.0.1
fastapi==0.95.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
packaging==23.0
pydantic==1.10.7
redis==4.2.2
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.19.0
wrapt==1.15.0
yarl==1.8.2
